+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2023-09-30 - emadriga
+### Changed
+- Grouped chat by dates chat messages with time
+
 ## 2023-09-28 - emadriga
 ### Changed
 - Functional chat containing websockets sharing minimal info ws to receptors only and just new info not old. Also displays connected users.